requirements.txt
py-cord
httpx
python-dotenv
aiosqlite
Pillow
